# minilog: remove data race