Add an RWMutex to minilog to synchronize access to global state. If you were adding a logger while something was logging, a concurrent map access could cause a panic.

This adds some overhead but it seems pretty minimal (from BenchmarkLogging):

Without lock: 77560 ns/op, total time: 3.022s
With lock: 78526 ns/op, total time: 3.052s

Fixes #515